Monkey patch CORS requests policy
A temporary solution for CORS policy not working randomly, this may be
an active bug either on the deployment service or laravel, will be fixed
later as we have to move on and work on more features as priority.
This monkey patch won't affect future code quality since CORS is a
middleware problem.

Signed-off-by: Yidir <yader.dev@gmail.com>